[*] Technical - Clean some resources services
