+++ src/services/resource-updater.js
@@ -1,39 +1,39 @@
-'use strict';
-var _ = require('lodash');
-var ErrorHTTP422 = require('./errors').ErrorHTTP422;
-var Interface = require('forest-express');
-var ResourceGetter = require('./resource-getter');
-var CompositeKeysManager = require('./composite-keys-manager');
-var ResourceFinder = require('./resource-finder');
+const _ = require('lodash');
+const { ErrorHTTP422 } = require('./errors');
+const Interface = require('forest-express');
+const ResourceGetter = require('./resource-getter');
+const CompositeKeysManager = require('./composite-keys-manager');
+const ResourceFinder = require('./resource-finder');
 
 function ResourceUpdater(model, params, newRecord) {
-  var schema = Interface.Schemas.schemas[model.name];
+  const schema = Interface.Schemas.schemas[model.name];
 
-  this.perform = function () {
-    var compositeKeysManager =
+  this.perform = () => {
+    const compositeKeysManager =
       new CompositeKeysManager(model, schema, newRecord);
 
     return new ResourceFinder(model, params)
       .perform()
-      .then(function (record) {
+      .then((record) => {
         if (record) {
-          _.each(newRecord, function (value, attribute) {
+          _.each(newRecord, (value, attribute) => {
             record[attribute] = value;
           });
 
           return record.validate()
-            .catch(function (error) { throw new ErrorHTTP422(error.message); })
-            .then(function () { return record.save(); });
+            .catch((error) => { throw new ErrorHTTP422(error.message); })
+            .then(() => record.save());
         }
+        return null;
       })
-      .then(function () {
+      .then(() => {
         if (schema.isCompositePrimary) {
           newRecord.forestCompositePrimary =
             compositeKeysManager.createCompositePrimary();
         }
 
         return new ResourceGetter(model, {
-          recordId: params.recordId
+          recordId: params.recordId,
         }).perform();
       });
   };
